Update the tests (#13)
* Add a bunch of gop tests and moved gs_test_old

Added tests for min, max and add; Moved src/gs_test_old to
tests/gs_test_old; Updated .travis.yml and Added a script to run the
tests in test directory. Also, polished the Makefile a bit as well.

* Fix the failing tests

findpts_el_2_test2.c:
When the test returns number of failures, it breaks Travis tests. So
I modified it to check if the number of failures are equal to the expected
value. If the test is not counting failures, we will simply return 0.

gs_test_old.c:
Convert C communicator to fortran and set datatype to 3 instead of 4.